From http://devl.org/pipermail/mc1322x/2012-July/001394.html
After some investigation, I found out that it is due to some memory
being used when it has been freed previously (and being freed again
afterward). The culprit is on line 248 of linux.c. The break statement
breaks the "for loop". However, I believe the intended behavior was to
break the switch-case.